Move an incorrectly bucketed test from tests/test_tracebacks.py.
I accidentally put a test with a type annotation in this file, which was
blocking a GitHub export.

PiperOrigin-RevId: 224854721
