+++ pytype/tests/py3/CMakeLists.txt
@@ -47,6 +47,15 @@ py_test(
 
 py_test(
   NAME
+    test_tracebacks
+  SRCS
+    test_tracebacks.py
+  DEPS
+    pytype.tests.test_base
+)
+
+py_test(
+  NAME
     test_cmp
   SRCS
     test_cmp.py
