+++ pytype/tests/test_tracebacks.py
@@ -68,13 +68,5 @@ class TracebackTest(test_base.TargetIndependentTest):
     """)
     self.assertErrorLogIs(errors, [(2, "attribute-error", r"attr.*None$")])
 
-  def test_build_class(self):
-    errors = self.CheckWithErrors("""\
-      class Foo(object):
-        def f(self, x: Bar):
-          pass
-    """)
-    self.assertErrorLogIs(errors, [(2, "name-error", r"Bar.*not defined$")])
-
 
 test_base.main(globals(), __name__ == "__main__")
